Look connected
Look fixed and working.
Ammo/Score not working in widget due to having the score and ammo functions built up on BP weapon/BP rifle for ammo, and BP Enemy as a parent for BP Ammo... Would have to convert/connect all of them with C++ to get the widget score/ammo to update accordingly.  In the background the score and ammo is actually functioning, just not connected to widget.